Fix color index mode rendering. Changed SWvertex's index field to GLfloat and fix a few other bits.

+++ src/mesa/swrast/swrast.h
@@ -1,8 +1,8 @@
 /*
  * Mesa 3-D graphics library
- * Version:  5.1
+ * Version:  6.1
  *
- * Copyright (C) 1999-2003  Brian Paul   All Rights Reserved.
+ * Copyright (C) 1999-2004  Brian Paul   All Rights Reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the "Software"),
@@ -68,7 +68,7 @@ typedef struct {
    GLchan color[4];
    GLchan specular[4];
    GLfloat fog;
-   GLuint index;
+   GLfloat index;
    GLfloat pointSize;
 } SWvertex;
